added BlackPowderWeaponScript class
